@@ -1,6 +1,6 @@
 class Song {
   final String title;
   final String duration;
-  final String path;
+  final String? path;
   const Song(this.title, this.duration, {this.path});
 }
