@@ -4,7 +4,7 @@ import 'package:flutter/material.dart';
 
 class MusicianPage extends StatelessWidget {
   final Musician musician;
-  const MusicianPage(this.musician, {Key key}) : super(key: key);
+  const MusicianPage(this.musician, {Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
