@@ -1,10 +1,12 @@
 import 'package:flutter/material.dart';
 
 class LinearGradientTween extends Tween<LinearGradient> {
-  LinearGradientTween({LinearGradient begin, LinearGradient end})
+  LinearGradientTween(
+      {required LinearGradient begin, required LinearGradient end})
       : super(begin: begin, end: end);
+
   @override
-  LinearGradient lerp(double t) => LinearGradient.lerp(begin, end, t);
+  LinearGradient lerp(double t) => LinearGradient.lerp(begin, end, t)!;
 }
 
 class RadialGradientTween extends Tween<RadialGradient> {
@@ -17,7 +19,7 @@ final Animatable<LinearGradient> background = TweenSequence<LinearGradient>([
           begin: LinearGradient(
             begin: Alignment.topRight,
             end: Alignment.bottomLeft,
-            colors: [Colors.lightBlue, Colors.blue[800]],
+            colors: [Colors.lightBlue, Colors.blue[800]!],
           ),
           end: linear1),
       weight: 1.0),
@@ -27,7 +29,7 @@ final Animatable<LinearGradient> background = TweenSequence<LinearGradient>([
         end: LinearGradient(
           begin: Alignment.bottomLeft,
           end: Alignment.topRight,
-          colors: [Colors.orange, Colors.orange[200]],
+          colors: [Colors.orange, Colors.orange[200]!],
         ),
       ),
       weight: 1.0),
@@ -36,7 +38,7 @@ final Animatable<LinearGradient> background = TweenSequence<LinearGradient>([
           begin: LinearGradient(
             begin: Alignment.bottomCenter,
             end: Alignment.topCenter,
-            colors: [Colors.orange[300], Colors.orange[600]],
+            colors: [Colors.orange[300]!, Colors.orange[600]!],
           ),
           end: linear1),
       weight: 1.0),
@@ -55,7 +57,7 @@ const linear1 = LinearGradient(
 final linear2 = LinearGradient(
   colors: [
     Colors.orange,
-    Colors.orange[900],
+    Colors.orange[900]!,
   ],
   begin: Alignment.topRight,
   end: Alignment.bottomLeft,
